requirements.txt
fastapi==0.128.4
uvicorn[standard]==0.40.0
python-multipart==0.0.22
pydantic==2.11.7
PyMuPDF==1.26.3
sentence-transformers==5.2.2
chromadb==0.6.3
streamlit==1.45.0
requests==2.32.3
python-dotenv==1.0.1
beautifulsoup4==4.12.3
lxml==5.3.0
openpyxl==3.1.5
